requirements.txt
torch==1.6.0
numpy==1.19.1
pytorch-lightning==1.1.2
hydra-core==1.0.3
